+++ components/index.ts
@@ -1,12 +1,12 @@
 export * from './AuthModal';
-export * from './Button';
+export * from './Common/Button';
 export * from './ErrorFallback';
-export * from './Form';
+export * from './Common/Form';
 export * from './LandingPage';
 export * from './ProductHunt';
 export * from './ProjectsPage/ProjectsList/ProjectsList';
 export * from './AddProjectForm';
 export * from './Tags';
-export * from './Typography';
+export * from './Common/Typography';
 export * from './Toaster';
 export * from './Layouts';
